Created a prompt_templates.py file and moved prompt templates to config folder. Added new unit tests and renamed sample tests.
